fix: change bigint to int

+++ db/query/professor.sql
@@ -234,7 +234,7 @@ UPDATE professors
 SET
   status = @status
 WHERE
-  id = @id::bigint
+  id = @id::int
 RETURNING *;
 
 -- name: RandomProfessorID :one
