+++ db/query/school.sql
@@ -82,7 +82,7 @@ UPDATE schools
 SET
   status = @status
 WHERE
-  id = @id::bigint
+  id = @id::int
 RETURNING *;
 
 -- name: RandomSchoolID :one
